requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0
pydantic>=2.0

# LLM Providers
openai>=1.0
anthropic>=0.30
google-generativeai>=0.7
httpx>=0.27
ollama>=0.3

# Search
tavily-python>=0.4
serpapi>=0.1
duckduckgo-search>=6.0
beautifulsoup4>=4.12
requests>=2.31
aiohttp>=3.9

# RAG / Embeddings
chromadb>=0.5
sentence-transformers>=3.0

# UI / Output
rich>=13.0

# Telegram
python-telegram-bot>=21.0

# Sandbox
docker>=7.0

# Async I/O
aiofiles>=24.0
